LLM my own/chat bot/requirements.txt
langchain
streamlit
huggingface_hub
tiktoken
python-dotenv
faiss-cpu
transformers
